+++ src/mesh/model/mesh-information-element-vector.cc
@@ -16,9 +16,8 @@
 #include "ns3/ie-dot11s-beacon-timing.h"
 #include "ns3/ie-dot11s-configuration.h"
 #include "ns3/ie-dot11s-id.h"
+#include "ns3/ie-dot11s-mesh-peering-management.h"
 #include "ns3/ie-dot11s-metric-report.h"
-#include "ns3/ie-dot11s-peer-management.h"
-#include "ns3/ie-dot11s-peering-protocol.h"
 #include "ns3/ie-dot11s-perr.h"
 #include "ns3/ie-dot11s-prep.h"
 #include "ns3/ie-dot11s-preq.h"
@@ -117,7 +116,7 @@ MeshInformationElementVector::DeserializeSingleIe(Buffer::Iterator start)
         newElement = Create<dot11s::IeLinkMetricReport>();
         break;
     case IE_MESH_PEERING_MANAGEMENT:
-        newElement = Create<dot11s::IePeerManagement>();
+        newElement = Create<dot11s::IeMeshPeeringManagement>();
         break;
     case IE_BEACON_TIMING:
         newElement = Create<dot11s::IeBeaconTiming>();
@@ -133,9 +132,6 @@ MeshInformationElementVector::DeserializeSingleIe(Buffer::Iterator start)
         break;
     case IE_PERR:
         newElement = Create<dot11s::IePerr>();
-        break;
-    case IE11S_MESH_PEERING_PROTOCOL_VERSION:
-        newElement = Create<dot11s::IePeeringProtocol>();
         break;
     default:
         NS_FATAL_ERROR("Information element " << +id << " is not implemented");