+++ src/mesh/model/mesh-information-element-vector.h
@@ -15,10 +15,6 @@
 namespace ns3
 {
 
-#define IE11S_MESH_PEERING_PROTOCOL_VERSION                                                        \
-    ((WifiInformationElementId)74) // to be removed (Protocol ID should be part of the Mesh Peering
-                                   // Management IE)
-
 /**
  * @brief Information element vector
  * @ingroup wifi